created share custome hook
